tests/wpt/tests/css/cssom/cssom-cssstyledeclaration-set.html

<!DOCTYPE html>
<html>
    <head>
        <title>CSSOM: CSSStyleDeclaration on HTMLElement represents inline style changes</title>
        <link rel="author" title="Paul Irish" href="mailto:paul.irish@gmail.com">
        <link rel="help" href="http://www.w3.org/TR/cssom-1/#the-cssstyledeclaration-interface">

        <link rel="source" href="http://trac.webkit.org/export/120528/trunk/LayoutTests/fast/css/cssText-cache.html">
        <meta name="flags" content="dom">

        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
    </head>

    <body>
        <div id="log"></div>

        <div id="box"></div>

        <script>

          var style = document.getElementById('box').style;

          test(function(){

            style.left = "10px";
            assert_equals(style.left, "10px", 'left property set on element\'s CSSStyleDeclaration Object');
            style.left = "20px";
            assert_equals(style.left, "20px", 'left property set on element\'s CSSStyleDeclaration Object');

          }, 'CSSStyleDeclaration on HTMLElement represents inline style changes');

        </script>


    </body>
</html>
